+++ changelog.md
@@ -10,10 +10,9 @@ Project contributors:
 - Kuanysh
 - Jonathan
 
-## [v0.1.0] - Oct 6, 10.56 pm
+## [Unreleased] - Oct 5, 8.12 pm
 ### Updated
-- Fixed date sorting bug.
-- Updated indentation accross application.
+- Added info/fleet and info/flights controllers that display JSON data returned from models
 
 ## [Unreleased] - Oct 5, 7.56 pm
 ### Updated